Suppress error message for .notile and .timeout files not being valid images

+++ convert_and_compress_tiles.py
@@ -19,13 +19,14 @@ def main(indir):
     print('We would love to tell you how long this will take, but we cannot')
     for image_file in image_files:
         (image_filename, image_ext) = os.path.splitext(image_file)
-        try:
-            im = Image.open(image_file)
-            im.convert('YCbCr').save(image_filename + '.jpeg', 'JPEG', quality=70)
-            os.remove(image_file)
-        except:
-            print('{} is not a valid image file.'.format(image_file))
-        
+        if(image_ext != '.notile' and image_ext != '.timeout'):
+            try:
+                im = Image.open(image_file)
+                im.convert('YCbCr').save(image_filename + '.jpeg', 'JPEG', quality=70)
+                os.remove(image_file)
+            except:
+                print('{} is not a valid image file.'.format(image_file))
+            
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
